Page items bring to front and send to back - third pass, set z values #244

+++ mainApp/metatypes.h
@@ -403,7 +403,7 @@ public:
 class SceneObjectData
 {
 public:
-  float z;
+  qreal z;
   float scenePos[2];
   SceneObjectDirection direction;
   bool  armed;
